Use HTTP GET instead of POST. Rename busline_controller and change search results message
