+++ templates/search_result_page.html
@@ -7,7 +7,7 @@
         <h1 class="text-center"> Resultado</h1>
     </div>
     <div class = "row">
-        <h1 class="text-center"> Linhas encontradas: {{ count_busline }} </h1>
+        <h4 class="text-center"> Foram encontrados '{{ count_busline }}' linhas que possuem '{{ searched_number }}' em seus números. </h4>
     </div>
     <div class="row">
         <div class="small-12 large-12 columns">
