+++ templates/topbar.html
@@ -16,7 +16,7 @@
 		{% endif %}
 			<ul class="right">	
 				<li class="has-form">
-                 <form action="/buscar_linha/" method="POST">
+                 <form action="/buscar_linha/" method="GET">
                   {% csrf_token %}
     				  <div class="row collapse">
     				    <div class="large-8 small-9 medium-6 columns">
